+++ crates/turbopack/src/resolve_options_context.rs
@@ -43,6 +43,8 @@ pub struct ResolveOptionsContext {
     /// A list of rules to use a different resolve option context for certain
     /// context paths. The first matching is used.
     pub rules: Vec<(ContextCondition, ResolveOptionsContextVc)>,
+    /// A list of plugins which get applied before (in the future) and after
+    /// resolving.
     pub plugins: Vec<ResolvePluginVc>,
     pub placeholder_for_future_extensions: (),
 }
